First Commit
Started somewhat of an "empty" shell to work on this project, I've yet
to fully understand how to use doxygen so expect a doxygen commit soon.

And yes, I talk as if I had a giant team of collaboratos on my commit
messages but it's more like I'm taking to my future self or my teacher
if he decides to review this git repo.