Finished I/O
You can see what we have left to do at TO_DO. ( I say we but it's me
alone, please help).

There is starting to be a huge difference bettween .hh files and .cc
files but I'll postpone it to after the big river functions are
implemented.

For future reference the steps should be:
1 - Create modular design (pure paper) & define naming conventions
2 - io_handler (writting down made up functions in hh with docs)
3 - implement big functions
4 - implement small functions

I'm pretty sure that's what I was trying at the begining tho, so I
probably failed at step 1.